requirements.txt
tensorflow==2.18.0
numpy==1.23.4
matplotlib==3.6.0
scikit-learn==1.1.3
pandas==1.5.2
Pillow==9.5.0
